moved menu to new new main folder, created MenuItem and a test class
