+++ src/main/com/app/controller/Menu.java
@@ -1,4 +1,4 @@
-package Menu;
+package com.app.controller;
 
 import java.util.Scanner;
