+++ src/main/com/app/model/MenuItem.java
@@ -1,0 +1,7 @@
+package com.app.model;
+
+interface MenuItem {
+
+ void run();
+
+}
