+++ src/main/com/app/model/createOrderMenuItem.java
@@ -1,0 +1,10 @@
+package com.app.model;
+
+public class createOrderMenuItem implements MenuItem{
+
+  @Override
+  public void run() {
+
+  }
+
+}
